pyproject.toml
[tool.black]
line-length=79
skip-magic-trailing-comma=true

[tool.poetry]
name = "ivycraft"
version = "0"
description = ""
authors = ["CircuitSacul <circuitsacul@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
hikari = "^2.0.0-alpha.108"
hikari-crescent = "^0.1.0-beta.9"
apgorm = "^1.0.0-beta.10"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.942"
nox = "^2022.1.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
